canvas rework, the canvas context is now private
