+++ demo/public/presets/images.json
@@ -104,6 +104,15 @@
       "random": false,
       "value": 0.5
     },
+    "rotate": {
+      "animation": {
+        "enable": true,
+        "speed": 5,
+        "sync": false
+      },
+      "random": true,
+      "value": 0
+    },
     "shape": {
       "character": {
         "fill": false,
